+++ src/types/newsletters.ts
@@ -4,7 +4,6 @@ export const NewsletterSchema = z.object({
   id: z.string(),
   name: z.string(),
   description: z.string().nullable(),
-  owner: z.string().nullable(),
   createdAt: z.string().datetime({ offset: true }).nullable(),
   subscribers: z.number().int().nullable(),
   invite: z.string().nullable(),